fix dtype roialign_fwd

+++ include/miopen/miopen.h
@@ -8207,10 +8207,10 @@ MIOPEN_EXPORT miopenStatus_t miopenRoIAlignForward(miopenHandle_t handle,
                                                    const void* rois,
                                                    const miopenTensorDescriptor_t outputDesc,
                                                    void* output,
-                                                   const int32_t alignedHeight,
-                                                   const int32_t alignedWidth,
+                                                   const uint64_t alignedHeight,
+                                                   const uint64_t alignedWidth,
                                                    const float spatialScale,
-                                                   const int32_t samplingRatio,
+                                                   const int64_t samplingRatio,
                                                    const bool aligned);
 
 /*! @brief Execute a Region-of-Interest (RoI) Align backward layer
