+++ src/roialign.cpp
@@ -48,10 +48,10 @@ miopenStatus_t RoIAlignForward(Handle& handle,
                                ConstData_t rois,
                                const TensorDescriptor& outputDesc,
                                Data_t output,
-                               const int alignedHeight,
-                               const int alignedWidth,
+                               const uint64_t alignedHeight,
+                               const uint64_t alignedWidth,
                                const float spatialScale,
-                               const int samplingRatio,
+                               const int64_t samplingRatio,
                                const bool aligned)
 {
     const auto problem = roialign::FwdProblemDescription{
