+++ src/roialign_api.cpp
@@ -38,10 +38,10 @@ extern "C" miopenStatus_t miopenRoIAlignForward(miopenHandle_t handle,
                                                 const void* rois,
                                                 const miopenTensorDescriptor_t outputDesc,
                                                 void* output,
-                                                const int32_t alignedHeight,
-                                                const int32_t alignedWidth,
+                                                const uint64_t alignedHeight,
+                                                const uint64_t alignedWidth,
                                                 const float spatialScale,
-                                                const int32_t samplingRatio,
+                                                const int64_t samplingRatio,
                                                 const bool aligned)
 {
     MIOPEN_LOG_FUNCTION(inputDesc,
